LC-number-of-zero-filled-subarrays [Runtime: 3 ms; Memory: 61.7 MB]

+++ array/number-of-zero-filled-subarrays.java
@@ -1,6 +1,6 @@
 class Solution {
     public long zeroFilledSubarray(int[] nums) {
-        int  res = 0, streak = 0;
+        long  res = 0, streak = 0;
 
         for(int num: nums){
             streak = (num ==0)? streak +1: 0;
